fix: Prevents deleting non-existent messages
Ensures that the `deleteMessage` function only attempts
to delete messages if they actually exist. This prevents
potential errors when attempting to delete non-existent
message IDs.

+++ src/modules/notifiers/gotify.ts
@@ -30,7 +30,9 @@ export class GotifyNotifier extends AbstractNotifier {
   }
 
   override async deleteMessageImpl(id: string): Promise<void> {
-    await this.request.delete(`/message/${id}?token=${env.gotifyUserToken}`)
+    if (this.hasMessageIdImpl(id)) {
+      await this.request.delete(`/message/${id}?token=${env.gotifyUserToken}`)
+    }
   }
 
   override async deleteAllMessagesImpl(): Promise<void> {
